Replace InterfaceDispatchMapIndirectionNode with EmbeddedPointerIndirectionNode
InterfaceDispatchMapIndirectionNode was mostly boiler plate code. Now we
have a mechanism for pointer indirection nodes, migrate the interface
map pointer nodes over to use it.

To make the EmbeddedPointerIndirectionNode work for dispatch maps, add a callback delegate invoked when the indirection node is added to the parent ArrayOfEmbeddedPointersNode. This allows the dispatch map's EEType to record the map's index in the DispatchMapTable array.