Complete set of my R2R bring-up changes
This is the complete set of my R2R bring-up changes relative to
Simon's latest check-in in the R2R branch (runnable Main). The
biggest amount of noise is caused by the propagation of tokens;
Simon is working on a cleaner solution for this by improving
the JIT - ILCompiler interface.

Thanks

Tomas