Make import sections optional
Based on Michal's advice I have refactored section table
manipulation so that we no longer need explicit indices
and so we can make the tables optional.

1) In ReadyToRunObjectWriter.cs, I have removed part of
the ugliness of populating the entrypoint table after
the graph calculation has finished. I have moved this bit
to ReadyToRunCodegenNodeFactory to the point where we're
constructing a MethodCodeNode.

The ugliness still remains for the EETypeNodes - I haven't
done anything about that as I assume that this is temporary,
according to Michal at some point we'll need to switch
over to a different type node representation.

2) In DelayLoadHelper.cs, I added some checks for
relocsOnly - this is mostly required so that we don't
need the IndexFromBeginningArray too early but in general
it's a nice perf optimization - we don't need to emit the
actual code when we're just collecting dependency info.

3) In Import.cs I had to slightly modify name formatting
as Table (ImportSectionNode) no longer exposes the explicit
index. Instead of changing it to use IndexFromBeginningOfArray
I switched it over to use the table name which is more
expressive.

4) In ImportSectionNode, I removed explicit section index
manipulation and I made emission of the data optional,
when there are no imports in the section we can skip
the node completely.

5) In ImportSectionTableNode.cs I have removed the homebrew
implementation of GetElementDataForNodes as unneeded and
incomplete - it wasn't initializing IndexFromBeginningOfArray.

6) In MethodEntryPointTableNode I had to refactor GC info
manipulation - as we're now adding methods to the entrypoint
table "early", we don't yet have the fixup info available
as it now uses IndexFromBeginningOfArray which is only
initialized once the nodes are emitted.

I resolved this by modifying EntryPoint so that it holds
a reference to the primary method node and emits the fixup
info "lazily" in GetData. I think it simplified the code
overall.

7) In RuntimeFunctionsTableNode.cs I hit a similar problem
regarding early unification of method GC info. I have
modified this by creating a new embedded node type
MethodGCInfoNode and using it to emit GC info lazily.

Thanks

Tomas