First passing R2R test involving generics
With the latest delta I'm passing the following trivial test:

    private static bool ConstructListOfInt()
    {
        List<int> listOfInt = new List<int>();
        return listOfInt.Count == 0;
    }

1) In MultiFileCompilationModuleGroup, I had to modify handling
of instantiated types to check the generic type definition,
otherwise the compiler identified List`1 as belonging to the current
compilation group and died while trying to compile it - this is
why I was seeing it request a call to a MethodDef, that was the bit

  .method /*0600421D*/ public hidebysig specialname rtspecialname
          instance void  .ctor() cil managed
  {
    // Code size       20 (0x14)
    .maxstack  8
    IL_0000:  ldarg.0
*** IL_0001:  call       instance void System.Object/*0200005A*/::.ctor() /* 06000183 */
    IL_0006:  nop
    IL_0007:  nop
    IL_0008:  ldarg.0
    IL_0009:  ldsfld     !0[] class System.Collections.Generic.List`1/*020005B8*/<!T>/*1B00029D*/::s_emptyArray /* 0A000A86 */
    IL_000e:  stfld      !0[] class System.Collections.Generic.List`1/*020005B8*/<!T>/*1B00029D*/::_items /* 0A000A87 */
    IL_0013:  ret
  } // end of method List`1::.ctor

in the List ctor. I don't know if my change is correct, please
let me know if the underlying issue should be fixed in a different way.

2) In ReadyToRunCodegenCompilation, I used the improved check per (1)
to make sure we don't try to compile methods like List`1..ctor.
At least I assume List simply resides in a different version bubble and
it cannot be compiled into the app unless we proclaim it non-versionable.

3) I have generalized the MethodImportSignature to MethodFixupSignature
similar to TypeFixupSignature to cover the common case where the signature
consists of the fixup kind ID and the MethodRef RID.

Thanks

Tomas