Remove MethodEntrypoint noise
With the Jit interface now able to handle ready-to-run compilation through `#if` guards, introduce `ReadyToRunCodegenNodeFactory.MethodEntrypointWithToken` which allows us to handle MethodRefs without polluting all the existing `MethodEntrypoint` calls.

In order to unify existing calls to `MethodEntrypoint` that are needed (ie, rooting the Main method) with `MethodEntrypointWithToken` we share the same node cache but with an empty mdToken for calls through `MethodEntrypoint`.

This change is currently untested other than verifying we can build the unit test app and get a reasonable looking output. I can't even verify it in r2rdump because that goes into an infinite loop parsing some GCInfo.