Generate debug info for r2r images
Build the debug info table when generating a ready-to-run binary. This contains variable / and native to IL offset information.

The JIT calls setVars / setBoundaries to provide both sets of information. These methods are now split for RyuJit / ReadyToRun compilation modes since we use the data differently in each. For instance, in RyuJit mode, we map the native code back to sequence points and directly back to line numbers. In ReadyToRun, we use the raw debug information.

The types `DebugVarInfo` and `DebugLocInfo` are used to transfer this information.