Remove existing R2R sections from output image
This change is not the right direction. It specifically filters out existing ready-to-run sections which are found by the R2R header table, listing their RVAs and sizes. However, after each section there is padding before the next thing in that section. We don't know what thing it is and what its alignment requirements are.

Thus, the correct way to strip existing artifacts is to copy over what we know about: IL blob, metadata, etc.